chore: login and register seperated
